Added a mongo Dockerfile to prepare for storing orders of products in mongodb. Created custom annotation so JaxRs resources has a @HttpEndpoint annotation instead of @Controller
