+++ rest-services/src/main/java/com/github/andtho/config/HttpEnpoint.java
@@ -1,0 +1,7 @@
+package com.github.andtho.config;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public @interface HttpEnpoint {
+}
