--- rest-services/src/test/java/com/github/andtho/AbstractTestbase.java
@@ -1,4 +1,0 @@
-package com.github.andtho;
-
-public abstract class AbstractTestbase {
-}
